silenced compiler warnings and changed C casts to C++ casts
git-svn-id: svn+ssh://svn.pointclouds.org/pcl/branches/pcl-1.x@4914 a9d63959-f2ad-4865-b262-bf0e56cfafb6

+++ search/include/pcl/search/search.h
@@ -33,7 +33,7 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  *
- * $Id: search.h 4502 2012-02-17 00:44:05Z gedikli $
+ * $Id: search.h 4826 2012-02-28 21:33:11Z bouffa $
  *
  */
 
@@ -195,7 +195,7 @@ namespace pcl
         nearestKSearch (const PointCloud &cloud, int index, int k,
                         std::vector<int> &k_indices, std::vector<float> &k_sqr_distances) const
         {
-          assert (index >= 0 && index < (int)cloud.points.size () && "Out-of-bounds error in nearestKSearch!");
+          assert (index >= 0 && index < static_cast<int> (cloud.points.size ()) && "Out-of-bounds error in nearestKSearch!");
           return (nearestKSearch (cloud.points[index], k, k_indices, k_sqr_distances));
         }
 
@@ -222,13 +222,13 @@ namespace pcl
         {
           if (indices_ == NULL)
           {
-            assert (index >= 0 && index < (int)input_->points.size () && "Out-of-bounds error in nearestKSearch!");
+            assert (index >= 0 && index < static_cast<int> (input_->points.size ()) && "Out-of-bounds error in nearestKSearch!");
             return (nearestKSearch (input_->points[index], k, k_indices, k_sqr_distances));
           }
           else
           {
-            assert (index >= 0 && index < (int)indices_->size () && "Out-of-bounds error in nearestKSearch!");
-            if (index >= (int)indices_->size() || index < 0)
+            assert (index >= 0 && index < static_cast<int> (indices_->size ()) && "Out-of-bounds error in nearestKSearch!");
+            if (index >= static_cast<int> (indices_->size ()) || index < 0)
               return (0);
             return (nearestKSearch (input_->points[(*indices_)[index]], k, k_indices, k_sqr_distances));
           }
@@ -250,14 +250,14 @@ namespace pcl
             k_indices.resize (cloud.size ());
             k_sqr_distances.resize (cloud.size ());
             for (size_t i = 0; i < cloud.size (); i++)
-              nearestKSearch (cloud,(int) i,k,k_indices[i],k_sqr_distances[i]);
+              nearestKSearch (cloud, static_cast<int> (i), k, k_indices[i], k_sqr_distances[i]);
           }
           else
           {
             k_indices.resize (indices.size ());
             k_sqr_distances.resize (indices.size ());
             for (size_t i = 0; i < indices.size (); i++)
-              nearestKSearch (cloud,indices[i],k,k_indices[i],k_sqr_distances[i]);
+              nearestKSearch (cloud, indices[i], k, k_indices[i], k_sqr_distances[i]);
           }
         }
 
@@ -360,7 +360,7 @@ namespace pcl
                       std::vector<int> &k_indices, std::vector<float> &k_sqr_distances,
                       unsigned int max_nn = 0) const
         {
-          assert (index >= 0 && index < (int)cloud.points.size () && "Out-of-bounds error in radiusSearch!");
+          assert (index >= 0 && index < static_cast<int> (cloud.points.size ()) && "Out-of-bounds error in radiusSearch!");
           return (radiusSearch(cloud.points[index], radius, k_indices, k_sqr_distances, max_nn));
         }
 
@@ -389,12 +389,12 @@ namespace pcl
         {
           if (indices_ == NULL)
           {
-            assert (index >= 0 && index < (int)input_->points.size () && "Out-of-bounds error in radiusSearch!");
+            assert (index >= 0 && index < static_cast<int> (input_->points.size ()) && "Out-of-bounds error in radiusSearch!");
             return (radiusSearch (input_->points[index], radius, k_indices, k_sqr_distances, max_nn));
           }
           else
           {
-            assert (index >= 0 && index < (int)indices_->size () && "Out-of-bounds error in radiusSearch!");
+            assert (index >= 0 && index < static_cast<int> (indices_->size ()) && "Out-of-bounds error in radiusSearch!");
             return (radiusSearch (input_->points[(*indices_)[index]], radius, k_indices, k_sqr_distances, max_nn));
           }
         }
@@ -422,7 +422,7 @@ namespace pcl
             k_indices.resize (cloud.size ());
             k_sqr_distances.resize (cloud.size ());
             for (size_t i = 0; i < cloud.size (); i++)
-              radiusSearch (cloud,(int) i,radius,k_indices[i],k_sqr_distances[i], max_nn);
+              radiusSearch (cloud, static_cast<int> (i), radius,k_indices[i], k_sqr_distances[i], max_nn);
           }
           else
           {
@@ -505,7 +505,7 @@ namespace pcl
     {
       std::vector<int> order (indices.size ());
       for (size_t idx = 0; idx < order.size (); ++idx)
-        order [idx] = idx;
+        order [idx] = static_cast<int> (idx);
 
       Compare compare (distances);
       sort (order.begin (), order.end (), compare);
